+++ src/AbilitySystem/ItemAbilities/RockSpike.cpp
@@ -36,7 +36,7 @@ void RockSpike::update(double timestep)
     if (currentTime <= 0.0)
     {
         currentTime = period;
-        GameModel::getInstance()->getResourcesModel()->addResource(Enums::ResourceTypes::STONE, amount);
+        GameModel::getInstance()->getResourcesModel()->addResource(static_cast<int>(Enums::ResourceTypes::STONE), amount);
     }
     else
         currentTime -= timestep;
